Add Conteo items - Consulta favs

+++ src/main/java/com/meli/mla/configuration/dto/StatsDTO.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class StatsDTO {
 
     private String id;
-    private Double quantity;
+    private Long quantity;
 }
